+++ back/src/main/java/com/openclassrooms/mddapi/repositories/PostsRepository.java
@@ -3,5 +3,5 @@ package com.openclassrooms.mddapi.repositories;
 import com.openclassrooms.mddapi.models.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostsRepository extends JpaRepository<Post, Integer> {
+public interface PostsRepository extends JpaRepository<Post, Long> {
 }
